+++ src/features/coins/coinsSlice.ts
@@ -1,24 +1,12 @@
 import { createSlice, createAsyncThunk } from "@reduxjs/toolkit";
 import type { PayloadAction } from "@reduxjs/toolkit";
 import { initialStateProp, globalStatsProp, singleCoinProp, coinsProp } from "../../interface";
-const baserUrl = 'https://coinranking1.p.rapidapi.com/'
-const options = {
-	method: 'GET',
-	headers: {
-    'x-rapidapi-key':'f0021db587msh781fb1cbef39856p11c183jsn45521d5d1c85',
-		// 'X-RapidAPI-Key': '2006fcddebmshb8e2fb348e572f5p13c9acjsn19e6f039cdde',
-		'X-RapidAPI-Host': 'coinranking1.p.rapidapi.com'
-	}
-};
-
+import { fetchCoins } from "../../api/coinsApi";
 export const getCoins = createAsyncThunk('coins/getCoins', async (count: number) => {
   try {
-    const response = await fetch(`${baserUrl}coins?limit=${count}`, options);
-    const result = await response.json();
-    console.log(result);
+    const result = await fetchCoins(count);
     return result;
   } catch (error) {
-    console.error(error);
     return error;
   }
 })
